Refactor pref_labels as standalone script
